requirements.txt
openai
openai-agents
pydantic
pyyaml
restate_sdk[serde]
hypercorn
